Fix: Remembers variables from previous calls
Fix to issue #99: "Do we need to provide all variables with cssVars.variables?".

Previously, calling cssVars with a new set of variables would reset any variables set in a previous call. That meant you had to include every variable with every call. Now, it uses the previous set of variables as a base for any new variables that are changing, so you only have to send the variables that have changed since the last call.